Expand all CTS with cmdlists to use both RCL and ICL (6/N)
Related-To: VLCLJ-1882

+++ conformance_tests/core/test_memory_overcommit/kernels/test_fill_device_memory_overcommit.cl
@@ -1,36 +1,36 @@
 /*
  *
- * Copyright (C) 2019 Intel Corporation
+ * Copyright (C) 2019-2023 Intel Corporation
  *
  * SPDX-License-Identifier: MIT
  *
  */
 
-typedef unsigned long uint64_t;
-typedef unsigned uint32_t;
-typedef unsigned short uint16_t;
+typedef ulong uint64_t;
+typedef uint uint32_t;
+typedef ushort uint16_t;
 
 kernel void fill_device_memory(__global uint64_t *pattern_memory,
-                               const size_t pattern_memory_count,
+                               const uint64_t pattern_memory_count,
                                const uint16_t pattern_base) {
   uint32_t i;
 
-  /* fill memory with pattern */
+  /* Fill memory with pattern */
   for (i = 0; i < pattern_memory_count; i++)
     pattern_memory[i] = (i << (sizeof(uint16_t) * 8)) + pattern_base;
 }
 
 /*
  * Verify pattern buffer against expected pattern.
- * In case of unexpected differences, use output buffers to record
- * some of those differences.
+ * In case of unexpected differences, use found_output buffers to record
+ * some of those differences and expected_output buffer to record golden data
  */
 kernel void test_device_memory(__global uint64_t *pattern_memory,
-                               const size_t pattern_memory_count,
+                               const uint64_t pattern_memory_count,
                                const uint16_t pattern_base,
                                __global uint64_t *expected_output,
                                __global uint64_t *found_output,
-                               const size_t output_count) {
+                               const uint32_t output_count) {
   uint32_t i, j;
 
   j = 0;
